Add static cast.

+++ fix/src/cpp/BridgeImpl.cpp
@@ -74,8 +74,8 @@ BridgeImpl::BridgeImpl(const string& name, const Ice::CommunicatorPtr& communica
     _timer(new IceUtil::Timer()),
     _retryInterval(IceUtil::Time::seconds(communicator->getProperties()->getPropertyAsIntWithDefault(
                                               name + ".RetryInterval", 60))),
-    _forwardTimeout(communicator->getProperties()->getPropertyAsIntWithDefault(
-                        name + ".RetryInterval", 60)/2.0 * 1000),
+    _forwardTimeout(static_cast<int>(communicator->getProperties()->getPropertyAsIntWithDefault(
+                                         name + ".RetryInterval", 60)/2.0 * 1000)),
     _initiator(0),
     _active(false),
     _session(0)
